+++ src/f32/vec3.rs
@@ -677,16 +677,28 @@ impl Vec3 {
         }
     }
 
-    /// Returns a vector containing the fractional part of the vector, e.g. `self -
-    /// self.trunc()`.
+    /// Returns a vector containing the fractional part of the vector as `self - self.trunc()`.
     ///
-    /// Note that this differs from the GLSL specification which uses `self - self.floor()`.
+    /// Note that this differs from the GLSL implementation of `fract` which returns
+    /// `self - self.floor()`.
     ///
     /// Note that this is fast but not precise for large numbers.
     #[inline]
     #[must_use]
     pub fn fract(self) -> Self {
         self - self.trunc()
+    }
+
+    /// Returns a vector containing the fractional part of the vector as `self - self.floor()`.
+    ///
+    /// Note that this differs from the Rust implementation of `fract` which returns
+    /// `self - self.trunc()`.
+    ///
+    /// Note that this is fast but not precise for large numbers.
+    #[inline]
+    #[must_use]
+    pub fn fract_gl(self) -> Self {
+        self - self.floor()
     }
 
     /// Returns a vector containing `e^self` (the exponential function) for each element of